// src/services/homes/types/home.ts
export type Home = {
  data: Data;
  status: number;
  message: string;
  cached: boolean;
};

export type Data = {
  name: string;
  description: any;
  slug: string;
  parentId: any;
  seo: Seo;
  homeFields: HomeFields;
  color: any;
  background: any;
};

export type Seo = {
  title: string;
  metaDesc: string;
  metaKeywords: any;
  metaRobotsNofollow: string;
  metaRobotsNoindex: string;
  canonical: string;
  opengraphTitle: string;
  opengraphDescription: string;
  opengraphImage: any;
  twitterTitle: string;
  twitterDescription: string;
  twitterImage: any;
};

export type HomeFields = {
  homeType: string;
  featuredBlocks5: FeaturedBlocks5;
  sliderBlocks: SliderBlock[];
  sliderSecondaryBlocks: SliderSecondaryBlock[];
  embedBlocks: any;
};
// ...
export type SliderBlock = {
  sliderBlock: SliderBlock2;
};

export type SliderBlock2 = {
  featuredPost: FeaturedPost6[];
  custom: string;
  title: string;
  subtitle: string;
  url: string;
  image: Image12;
};

export type FeaturedPost6 = {
  title: string;
  uri: string;
  link: string;
  slug: string;
  primaryCatName: string;
  extraFields: ExtraFields6;
};

export type ExtraFields6 = {
  subtitle: string;
  image: Image11;
};

export type Image11 = {
  sourceUrl: string;
  altText: string;
  caption: any;
  description: any;
};

export type Image12 = {
  sourceUrl: string;
  altText: string;
  caption: any;
  description: any;
};